refactor: `formattedMeter` 패키지 변경
